Add rule to detect multiple content emissions in a composable (#9)
* Add rule to detect multiple content emissions in a composable

Composables that emit content from multiple composables at its top level are problematic: we don't know how they will be rendered.

For example:
```kotlin
@Composable
private fun InnerContent() {
    Text(...)
    Image(...)
    Button(...)
}
```

If this InnerContent composable was wrapped in a `Column`, it will stack them vertically, but if it were wrapped in a `Row`, it will show them horizontally. Instead, InnerContent should be cohesive and emit a single layout node itself.