Updates ktlint to 0.46.1
ktlint had a significant update in 0.46.x, including a bunch of breaking changes in tests. The new API is nicer though.
This patch includes all the migration to the new test style, and to pass the default tests.

NOTE: At the moment of writing this, spotless hasn't been updated yet to support the new ktlint version. https://github.com/diffplug/spotless/issues/1239