Separate modifier rules and update ordering checks
In this patch, I am separating the two modifier rules that were bndled in ComposeModifierMissingCheck. I moved the missing default modifier one to its own checker, ComposeModifierWithoutDefaultCheck, so it can be disabled individually if needed. They also don't have the same requirements, so this is fair.
When looking at the docs to write the rules.md blurb, I noticed there was an issue in the ordering rule related to modifier names, so I added this to the patch as well as it's all about fixing modifier rules.