Migrate to ktlint 0.47.0 (#24)
* Migrate to ktlint 0.47.0

There are a bunch of [breaking API changes](https://github.com/pinterest/ktlint/releases/tag/0.47.0) in the newest version of ktlint. Basically all rulesets need to be redefined. In 0.47.0 versions the previous ruleset versions coexist with the newers, so I added support for both in the mean time. In the future, the previous version of the rulesets will disappear. Also, undocumented by the migration guide, but all tests changed the way in which the assertThat classes are created, so I also updated all the tests to support the new version.

Spotless is not yet compatible with 0.47.0, so I disabled the checks until it is (will likely [land soon](https://github.com/diffplug/spotless/pull/1303)).

I also updated all dependencies.