Add infra to auto publish the library in CI
I am adding the ability to automatically release artifacts into Maven Central for non snapshot versions. For snapshots, this step will be ignored - for that to happen I needed to make the VERSION from gradle.properties visible, which was easy to do by registering a new gradle task for it.