Make the namings consistent for Detekt (#38)
Detekt favors pascal case for names of rulesets and rules, and we were using the same names as the ktlint rules. This patch changes the name of the custom ruleset to `TwitterCompose`, and all the individual rules to `PascalCase`.
I also added the list of all the individual rules to enable for `detekt.yml` ready for copypasta, just to make it easier for folks that want to use these.