Added shadowJar to be able to create farjars for ktlint/detekt rules (#65)
Using `shadowJar`, we can now generate fat jars (that contain all the dependencies necessary in them) for both `:rules:ktlint` and `:rules:detekt`. This is important to be able to run the rules via CLI in both of them.

Tested locally for both CLIs and worked fine. Folks, remember to have a detekt.yml allowlisting all the rules, like described in the docs :smile: