Update docs with ktlint/detekt instructions and new uber jars
Using the rules with ktlint is an adventure, and for now only kotlinter seems a valid option. I have kept the ktlint-gradle and spotless sections as a FAQ though.
I've also added mentions to use the uber jars added to releases for ktlint/detekt CLI in the docs.

Given that the "Overview" section was getting a bit big, I've broken down in two separate pages for ktlint and detekt.

Supersedes #54 by adding extra info.