Add ComposeCompositionLocalAllowlist rule (#80)
Adds a rule to disallow the introduction of new `CompositionLocal`s, and adds a `allowedCompositionLocals` (Detekt) and `allowed_composition_locals` (ktlint) config entries to be able to provide an allowed list of them (e.g. if they are necessary for some infra or other valid cases like that, don't go abusing this lol).