Add twitter_compose_ prefix to ktlint rules config entries
This patch changes how the rules are named in ktlint, to add a prefix to the rule (which is a common practice, as all config is bundled together which makes it hard to contextualize what it configures). Detekt configs are unchanged, as they are well scoped out of the box in its config files.